Each action has a sprite shown
Perfectly set up now for animations. Made a universal timeline that checks what inputs are made and it turns on the sprite image object on the canvas associated with it. I reorganized the way methods are called. Each action now has a "True" method which invokes either to be a called upon particle effect. Or if motion is needed a Timeline.
There's more files now due to the asset store giving me particle effects.  They are cool though I just need to find the best way to use them. I know how but things are very sequential at this point in the game logic so I gotta be careful ya feel.